Перенеси код в модуль stats.js, создай отдельный файл с тестами
